Rename header file.
