+++ py/vmentrytable.h
